refactor
- align `gix-hash` integration tests with 'new' style.
- reorganize `hasher` module to avoid duplicate paths to the same types/functions.
- use shorter paths to `gix_hash::hasher|io` everywhere.
