+++ gix-hash/src/lib.rs
@@ -15,8 +15,11 @@ pub use borrowed::{oid, Error};
 
 /// Hash functions and hash utilities
 pub mod hasher;
-pub use hasher::io::{bytes, bytes_of_file, bytes_with_hasher};
-pub use hasher::{hasher, Hasher};
+pub use hasher::_impl::{hasher, Hasher};
+
+/// Error types for utility hash functions
+pub mod io;
+pub use io::_impl::{bytes, bytes_of_file, bytes_with_hasher};
 
 mod object_id;
 pub use object_id::{decode, ObjectId};
