+++ gix-hash/tests/hash/kind.rs
