+++ gix-hash/tests/hash/main.rs
