+++ gix-index/src/file/write.rs
@@ -1,5 +1,3 @@
-use gix_hash::hasher;
-
 use crate::{write, File, Version};
 
 /// The error produced by [`File::write()`].
@@ -7,7 +5,7 @@ use crate::{write, File, Version};
 #[allow(missing_docs)]
 pub enum Error {
     #[error(transparent)]
-    Io(#[from] hasher::io::Error),
+    Io(#[from] gix_hash::io::Error),
     #[error("Could not acquire lock for index file")]
     AcquireLock(#[from] gix_lock::acquire::Error),
     #[error("Could not commit lock for index file")]
@@ -21,14 +19,14 @@ impl File {
         &self,
         mut out: impl std::io::Write,
         options: write::Options,
-    ) -> Result<(Version, gix_hash::ObjectId), hasher::io::Error> {
+    ) -> Result<(Version, gix_hash::ObjectId), gix_hash::io::Error> {
         let _span = gix_features::trace::detail!("gix_index::File::write_to()", skip_hash = options.skip_hash);
         let (version, hash) = if options.skip_hash {
             let out: &mut dyn std::io::Write = &mut out;
             let version = self.state.write_to(out, options)?;
             (version, self.state.object_hash.null())
         } else {
-            let mut hasher = hasher::io::Write::new(&mut out, self.state.object_hash);
+            let mut hasher = gix_hash::io::Write::new(&mut out, self.state.object_hash);
             let out: &mut dyn std::io::Write = &mut hasher;
             let version = self.state.write_to(out, options)?;
             (version, hasher.hash.try_finalize()?)
