+++ gix-object/benches/edit_tree.rs
@@ -136,10 +136,7 @@ criterion_main!(benches);
 
 type TreeStore = Rc<RefCell<gix_hashtable::HashMap<ObjectId, Tree>>>;
 
-fn new_inmemory_writes() -> (
-    TreeStore,
-    impl FnMut(&Tree) -> Result<ObjectId, gix_hash::hasher::io::Error>,
-) {
+fn new_inmemory_writes() -> (TreeStore, impl FnMut(&Tree) -> Result<ObjectId, gix_hash::io::Error>) {
     let store = TreeStore::default();
     let write_tree = {
         let store = store.clone();
