+++ gix-pack/src/index/write/mod.rs
@@ -2,7 +2,6 @@ use std::{io, sync::atomic::AtomicBool};
 
 pub use error::Error;
 use gix_features::progress::{self, prodash::DynNestedProgress, Count, Progress};
-use gix_hash::hasher;
 
 use crate::cache::delta::{traverse, Tree};
 
@@ -181,7 +180,7 @@ impl crate::index::File {
 
         root_progress.inc();
 
-        let (resolver, pack) = make_resolver().map_err(hasher::io::Error::from)?;
+        let (resolver, pack) = make_resolver().map_err(gix_hash::io::Error::from)?;
         let sorted_pack_offsets_by_oid = {
             let traverse::Outcome { roots, children } = tree.traverse(
                 resolver,
@@ -225,7 +224,7 @@ impl crate::index::File {
                 let header = crate::data::header::encode(pack_version, 0);
                 let mut hasher = gix_hash::hasher(object_hash);
                 hasher.update(&header);
-                hasher.try_finalize().map_err(hasher::io::Error::from)?
+                hasher.try_finalize().map_err(gix_hash::io::Error::from)?
             }
             None => return Err(Error::IteratorInvariantTrailer),
         };
@@ -256,7 +255,7 @@ fn modify_base(
     pack_entry: &crate::data::Entry,
     decompressed: &[u8],
     hash: gix_hash::Kind,
-) -> Result<(), hasher::Error> {
+) -> Result<(), gix_hash::hasher::Error> {
     let object_kind = pack_entry.header.as_kind().expect("base object as source of iteration");
     let id = gix_object::compute_hash(hash, object_kind, decompressed)?;
     entry.id = id;
