+++ gix-pack/src/multi_index/write.rs
@@ -5,7 +5,6 @@ use std::{
 };
 
 use gix_features::progress::{Count, DynNestedProgress, Progress};
-use gix_hash::hasher;
 
 use crate::multi_index;
 
@@ -15,7 +14,7 @@ mod error {
     #[allow(missing_docs)]
     pub enum Error {
         #[error(transparent)]
-        Io(#[from] gix_hash::hasher::io::Error),
+        Io(#[from] gix_hash::io::Error),
         #[error("Interrupted")]
         Interrupted,
         #[error(transparent)]
@@ -84,7 +83,7 @@ impl multi_index::File {
         should_interrupt: &AtomicBool,
         Options { object_hash }: Options,
     ) -> Result<Outcome, Error> {
-        let out = hasher::io::Write::new(out, object_hash);
+        let out = gix_hash::io::Write::new(out, object_hash);
         let (index_paths_sorted, index_filenames_sorted) = {
             index_paths.sort();
             let file_names = index_paths
@@ -183,7 +182,7 @@ impl multi_index::File {
             index_paths_sorted.len() as u32,
             object_hash,
         )
-        .map_err(hasher::io::Error::from)?;
+        .map_err(gix_hash::io::Error::from)?;
 
         {
             progress.set_name("Writing chunks".into());
@@ -191,7 +190,7 @@ impl multi_index::File {
 
             let mut chunk_write = cf
                 .into_write(&mut out, bytes_written)
-                .map_err(hasher::io::Error::from)?;
+                .map_err(gix_hash::io::Error::from)?;
             while let Some(chunk_to_write) = chunk_write.next_chunk() {
                 match chunk_to_write {
                     multi_index::chunk::index_names::ID => {
@@ -209,7 +208,7 @@ impl multi_index::File {
                     ),
                     unknown => unreachable!("BUG: forgot to implement chunk {:?}", std::str::from_utf8(&unknown)),
                 }
-                .map_err(hasher::io::Error::from)?;
+                .map_err(gix_hash::io::Error::from)?;
                 progress.inc();
                 if should_interrupt.load(Ordering::Relaxed) {
                     return Err(Error::Interrupted);
@@ -218,11 +217,11 @@ impl multi_index::File {
         }
 
         // write trailing checksum
-        let multi_index_checksum = out.inner.hash.try_finalize().map_err(hasher::io::Error::from)?;
+        let multi_index_checksum = out.inner.hash.try_finalize().map_err(gix_hash::io::Error::from)?;
         out.inner
             .inner
             .write_all(multi_index_checksum.as_slice())
-            .map_err(hasher::io::Error::from)?;
+            .map_err(gix_hash::io::Error::from)?;
         out.progress.show_throughput(write_start);
 
         Ok(Outcome { multi_index_checksum })
