+++ gix-pack/src/verify.rs
@@ -1,7 +1,6 @@
 use std::{path::Path, sync::atomic::AtomicBool};
 
 use gix_features::progress::Progress;
-use gix_hash::hasher;
 
 ///
 pub mod checksum {
@@ -45,10 +44,10 @@ pub fn checksum_on_disk_or_mmap(
         should_interrupt,
     ) {
         Ok(id) => id,
-        Err(hasher::io::Error::Io(err)) if err.kind() == std::io::ErrorKind::Interrupted => {
+        Err(gix_hash::io::Error::Io(err)) if err.kind() == std::io::ErrorKind::Interrupted => {
             return Err(checksum::Error::Interrupted);
         }
-        Err(hasher::io::Error::Io(_io_err)) => {
+        Err(gix_hash::io::Error::Io(_io_err)) => {
             let start = std::time::Instant::now();
             let mut hasher = gix_hash::hasher(object_hash);
             hasher.update(&data[..data_len_without_trailer]);
@@ -56,7 +55,7 @@ pub fn checksum_on_disk_or_mmap(
             progress.show_throughput(start);
             hasher.try_finalize()?
         }
-        Err(hasher::io::Error::Hasher(err)) => return Err(checksum::Error::Hasher(err)),
+        Err(gix_hash::io::Error::Hasher(err)) => return Err(checksum::Error::Hasher(err)),
     };
 
     actual.verify(&expected)?;
