+++ gix-status/src/index_as_worktree/types.rs
@@ -10,7 +10,7 @@ pub enum Error {
     #[error("The clock was off when reading file related metadata after updating a file on disk")]
     Time(#[from] std::time::SystemTimeError),
     #[error("IO error while writing blob or reading file metadata or changing filetype")]
-    Io(#[from] gix_hash::hasher::io::Error),
+    Io(#[from] gix_hash::io::Error),
     #[error("Failed to obtain blob from object database")]
     Find(#[from] gix_object::find::existing_object::Error),
     #[error("Could not determine status for submodule at '{rela_path}'")]
